Make agent callbacks optional + typecheck all examples on CI (#205)
## Description

Makes the agent callbacks optional.
Additionally, I noticed not all examples were typechecked on CI.
I added the checks + tidied up a bit.

## Motivation and Context

Simplifies the API and allow for additional callbacks in the future.

## Documentation impact

- [ ] Documentation update required
- [ ] Documentation updated [in another PR](_)
- [x] No documentation update required (docs are in the making)

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to
not work as expected) [the feature is not published yet, so we still can
have an api change]